OH MY GOD WE CAN HAVE INTEGERS

+++ game/src/api/AchievementProgress.hpp
@@ -14,7 +14,7 @@ namespace API {
 std::string toString() const;
 void fromString(const std::string &str);
         std::string achievement_id;
-        double progress;
+        int64_t progress;
         AchievementProgressType type = static_cast<AchievementProgressType>(0);
         AchievementProgressUserId user_id;
     };
